Add files via upload
The p-value associated with the logistic regression model is .190 and it is greater than .05 hence once again we failed to reject null hypotheses and unable to establish that landing page is statistically significant conversion. And the null hypothesis is now there is no difference between the treatment and control group. Alternative hypotheses are that there is a difference between the treatment and control groups. New hypotheses are: H0 : Pnew = Pold H1 : Pnew != Pold